Tormented Tactics

+++ src/main/java/net/eman3600/dndreams/items/magic_bow/MagicBowItem.java
@@ -93,7 +93,7 @@ public abstract class MagicBowItem extends BowItem {
         }
     }
 
-    public abstract float pullTime();
+    public abstract int pullTime();
 
     public abstract ItemStack getProjectile();
